+++ PhBuyModels/Orders.cs
@@ -7,7 +7,7 @@ namespace PhBuyModels
     {
         public decimal Id { get; set; }
         public decimal? ProductId { get; set; }
-        public decimal? Quantity { get; set; }
+        public short? Quantity { get; set; }
         public double? TotalPrice { get; set; }
         public string Status { get; set; }
         public decimal? SellerId { get; set; }
